Move namespace message to messages
