+++ components/loconetmessages/src/loconet/inboundhandler.hpp
@@ -1,14 +1,14 @@
 #ifndef LOCONET_INBOUNDHANDLER_HPP_
 #define LOCONET_INBOUNDHANDLER_HPP_
 
-#include "message/message.hpp"
+#include "messages/message.hpp"
 
 namespace loconet
 {
 	class InboundHandler
 	{
 	public:
-		virtual void handle_message(message::Message*);
+		virtual void handle_message(messages::Message*);
 	};
 }
